moved from_camel_case to user_relations.py
